Refactor chooseMatmulTile to Codegen dialect utils.
